+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   lots of memory
 - [Issue #437](https://github.com/influxdb/influxdb/issues/437). Queries with negative constants don't parse properly
 - [Issue #432](https://github.com/influxdb/influxdb/issues/432). Deleted data using a delete query is resurrected after a server restart
+- [Issue #439](https://github.com/influxdb/influxdb/issues/439). Report the right location of the error in the query
 
 ## v0.5.6 [2014-04-08]